+++ plugins/github-issues/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-github-issues
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.2
+  - @backstage/plugin-catalog-react@1.2.3
+
 ## 0.2.1
 
 ### Patch Changes