+++ plugins/newrelic-dashboard/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-newrelic-dashboard
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.2
+  - @backstage/plugin-catalog-react@1.2.3
+
 ## 0.2.5
 
 ### Patch Changes