+++ plugins/scaffolder/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-scaffolder
 
+## 1.9.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.2
+  - @backstage/integration-react@1.1.8
+  - @backstage/plugin-catalog-react@1.2.3
+
 ## 1.9.0
 
 ### Minor Changes